// src/common/enums/story-type.enum.ts
export enum StoryType {
  NEW_RELEASE = 'NEW_RELEASE',
  ALBUM = 'ALBUM',
  SINGLE = 'SINGLE',
  TOUR = 'TOUR',
  CONCERT = 'CONCERT',
  FESTIVAL = 'FESTIVAL',
  REUNION = 'REUNION',
  COLLABORATION = 'COLLABORATION',
  INTERVIEW = 'INTERVIEW',
  CONTROVERSY = 'CONTROVERSY',
  AWARD = 'AWARD',
  CHART = 'CHART',
  DEATH = 'DEATH',
  ANNOUNCEMENT = 'ANNOUNCEMENT',
  MUSIC_VIDEO = 'MUSIC_VIDEO',
  VIRAL = 'VIRAL',
  RUMOR = 'RUMOR',
  OTHER = 'OTHER',
}